Create table config file for machine job record tables / introduce table config principle
Commit creates a table config file for defining the base of the machine job record table.
This principle will be reused for other new and existing tables to reduce code duplication.

Adds:
- Table config principle
- Table config file for Machine Job Records
